arch2/storage/requirements.txt
flask
requests
grpcio==1.75.1
grpcio-tools==1.75.1
protobuf==6.32.1
